chore: rename FullPostContextProvider to FullPostContext
